fix: refactor nav stick logic
stupid window undefined error
